backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
google-genai>=1.0.0
python-dotenv==1.0.1
httpx==0.27.0
websockets>=13.0,<15.0
python-multipart==0.0.9
